+++ packages/models/src/user/mails.js
@@ -3,21 +3,23 @@
 const config = require('config').get('openSenseMap-API-models.integrations'),
   log = require('../log');
 
-const noMailerConfiguredFunc = function () {
-  return Promise.resolve({ 'msg': 'no mailer configured' });
+const noQueueConfiguredFunc = function () {
+  return Promise.resolve({ msg: 'no queue configured' });
 };
 
 module.exports = {
-  sendMail: noMailerConfiguredFunc
+  addToQueue: noQueueConfiguredFunc,
 };
 
-if (config.get('mailer.url')) {
+if (config.has('redis') && config.has('redis.host') && config.has('redis.port') && config.has('redis.username') && config.has('redis.password') && config.has('redis.db') && config.has('mailer.queue')) {
   /* eslint-disable global-require */
-  const got = require('got');
+  const { Queue } = require('bullmq');
   /* eslint-enable global-require */
 
+  let queue;
+
   const mailTemplates = {
-    'newBox' (user, box) {
+    newBox (user, box) {
       const sketchParams = {
         encoding: 'base64',
         ssid: '',
@@ -28,7 +30,7 @@ if (config.get('mailer.url')) {
         windSpeedPort: box.windSpeedPort,
         devEUI: '',
         appEUI: '',
-        appKey: ''
+        appKey: '',
       };
 
       if (box.access_token) {
@@ -37,64 +39,64 @@ if (config.get('mailer.url')) {
 
       return {
         payload: {
-          box
+          box,
         },
         attachment: {
           filename: 'senseBox.ino',
-          contents: box.getSketch(sketchParams)
-        }
+          contents: box.getSketch(sketchParams),
+        },
       };
     },
-    'newBoxLuftdaten' (user, box) {
+    newBoxLuftdaten (user, box) {
       return {
         payload: {
-          box
-        }
+          box,
+        },
       };
     },
-    'newBoxHackAir' (user, box) {
+    newBoxHackAir (user, box) {
       return {
         payload: {
-          box
-        }
+          box,
+        },
       };
     },
-    'newUser' (user) {
+    newUser (user) {
       return {
         payload: {
           token: user.emailConfirmationToken,
-          email: user.email
-        }
+          email: user.email,
+        },
       };
     }, // email confirmation request
-    'passwordReset' (user) {
+    passwordReset (user) {
       return {
         payload: {
-          token: user.resetPasswordToken
-        }
+          token: user.resetPasswordToken,
+        },
       };
     },
-    'newUserManagement' (user, boxes) {
+    newUserManagement (user, boxes) {
       return {
         payload: {
           boxes,
-          token: user.resetPasswordToken
-        }
+          token: user.resetPasswordToken,
+        },
       };
     },
-    'confirmEmail' (user) {
+    confirmEmail (user) {
       return {
         recipient: {
           address: user.unconfirmedEmail,
-          name: user.name
+          name: user.name,
         },
         payload: {
           token: user.emailConfirmationToken,
-          email: user.unconfirmedEmail
-        }
+          email: user.unconfirmedEmail,
+        },
       };
     },
-    'resendEmailConfirmation' (user) {
+    resendEmailConfirmation (user) {
       let addressToUse = user.email;
       if (user.unconfirmedEmail) {
         addressToUse = user.unconfirmedEmail;
@@ -103,73 +105,81 @@ if (config.get('mailer.url')) {
       return {
         recipient: {
           address: addressToUse,
-          name: user.name
+          name: user.name,
         },
         payload: {
           token: user.emailConfirmationToken,
-          email: addressToUse
-        }
+          email: addressToUse,
+        },
       };
     },
-    'newSketch' (user, box) {
+    newSketch (user, box) {
       return {
         payload: {
-          box
+          box,
         },
         attachment: {
           filename: 'senseBox.ino',
-          contents: box.getSketch({ encoding: 'base64', access_token: box.access_token })
-        }
+          contents: box.getSketch({
+            encoding: 'base64',
+            access_token: box.access_token,
+          }),
+        },
       };
     },
-    'deleteUser' () {
+    deleteUser () {
       return {};
-    }
+    },
   };
 
-  const requestMailer = (payload) => {
-    return got.post(config.get('mailer.url'), {
-      cert: config.get('cert'),
-      key: config.get('key'),
-      ca: config.get('ca_cert'),
-      json: payload,
-      ecdhCurve: 'auto'
-    })
-      .then((response) => {
-        log.info({ msg: 'successfully sent mails', mailer_response: response.body, mailer_response_code: response.statusCode });
+  const requestQueue = () => {
+    if (queue) {
+      return queue;
+    }
+    queue = new Queue(config.get('mailer.queue'), {
+      connection: {
+        host: config.get('redis.host'),
+        port: config.get('redis.port'),
+        username: config.get('redis.username'),
+        password: config.get('redis.password'),
+        db: config.get('redis.db'),
+      },
+    });
 
-        return response;
-      })
-      .catch((err) => {
-        throw err;
-      });
+    return queue;
   };
 
   module.exports = {
-    sendMail (template, user, data) {
-      if (!(template in mailTemplates)) {
-        return Promise.reject(new Error(`template ${template} not implemented`));
-      }
-
-      const { payload = {}, attachment, recipient = { address: user.email, name: user.name } } = mailTemplates[template](user, data);
+    addToQueue (template, user, data) {
+      const {
+        payload = {},
+        attachment,
+        recipient = { address: user.email, name: user.name },
+      } = mailTemplates[template](user, data);
 
       // add user and origin to payload
       payload.user = user;
       payload.origin = config.get('mailer.origin');
 
-      // complete the payload
-      const mailRequestPayload = [
-        {
-          template,
-          lang: user.language,
-          recipient,
-          payload,
-          attachment
-        }
-      ];
-
-      return requestMailer(mailRequestPayload);
-
+      return requestQueue().add(template, {
+        template,
+        lang: user.language,
+        recipient,
+        payload,
+        attachment
+      }, {
+        removeOnComplete: true,
+      })
+        .then((response) => {
+          log.info({
+            msg: 'Successfully added mail to queue',
+            job_id: response.id,
+            template: response.name
+          });
+        })
+        .catch((err) => {
+          throw err;
+        });
     }
   };
 }