+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # openSenseMap API Changelog
 
+## v11.1.0
+
+- Use @sensebox/opensensemap-api-models v3.1.0
+- New route `/users/me/boxes/:boxId` to get specific box of logged in user
+- New `page` parameter for `/users/me/boxes?page=<number>` for pagination of user embedded boxes. Default page size is 25.
+
+## v11.0.1
+
+- Use @sensebox/opensensemap-api-models v3.0.1
+
+## v11.0.0
+
+- Use @sensebox/opensensemap-api-models v3.0.0
+- Update Github Action dependencies
+- Update restify
+- Update Readme information
+
 ## v10.2.3
 
 - Use @sensebox/opensensemap-api-models v2.0.3