+++ packages/models/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## v3.1.0
+
+- Add `page` parameter to `getBoxes` method in user model
+- Add new `getBox` to user model for retreiving single box with secrets
+
+## v3.0.1
+
+- Update `@sensebox/sketch-templater` to fix compiling bug
+
+## v3.0.0
+
+- Upgrade `restify` to v9.0.0
+- Add SPS30 sensors
+- Remove old mailer and add Redis based queue system
+
 ## v2.0.3
 
 - Change handling `near` parameter (#694)
@@ -21,7 +36,7 @@
 - Added `sharedBoxes` functionality (#605)
 
 ## v1.3.1
-- Add SPS30 sensor 
+- Add SPS30 sensor
 - Update @sensebox/node-sketch-templater to v1.13.0
 
 ## v1.2.0